fix(renderer): achieve faithful flam3 parity — nthreads=1 goldens, final-xform, Double, op-order
Root cause of the parity gap was never Emberweft: the golden harness rendered with
flam3's default nthreads (=host CPUs), a multi-threaded sample split Emberweft's
single-threaded ISAAC port cannot match. flam3(nthreads=1) vs Emberweft is byte-identical.
Pin nthreads=1 in regen_goldens.sh (also makes goldens machine-independent).

Plus faithful-port fixes: Float->Double (match flam3 precision), operation-order
(palette*256, grid gutter, WHITE_LEVEL pre-scale, julia mult order, EPS, badvalue),
final-xform feedback (don't feed final-transformed point back into the trajectory —
matches flam3.c:275-296), and xform color default = index parity (flam3.c:3139).

All 6 golden genomes now pass strict >=30dB/>=0.95 (51-72 dB, SSIM≈1.0).

+++ Sources/FlameKit/Flam3Serializer.swift
@@ -80,7 +80,7 @@ public enum Flam3Serializer {
         "\(f6(t.a)) \(f6(t.b)) \(f6(t.c)) \(f6(t.d)) \(f6(t.e)) \(f6(t.f))"
     }
 
-    private static func f6(_ x: Float) -> String { String(format: "%.6f", x) }
+    private static func f6(_ x: Double) -> String { String(format: "%.6f", x) }
 
     private static func escape(_ s: String) -> String {
         s.replacingOccurrences(of: "&", with: "&amp;")
