+++ Sources/FlameReference/ReferenceRenderer.swift
@@ -19,7 +19,7 @@ public enum ReferenceRenderer {
             width: params.width, height: params.height, oversample: params.oversample,
             gamma: flame.quality.gamma, gammaThreshold: flame.quality.gammaThreshold,
             vibrancy: flame.quality.vibrancy,
-            sampleDensity: Float(params.samplesPerPixel),
+            sampleDensity: Double(params.samplesPerPixel),
             pixelsPerUnit: flame.camera.scale * pow(2, flame.camera.zoom))
     }
 }
